# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Vetfolio Manager API",
      "version": "1.0.0",
      "description": "API para la gestión de casos veterinarios"
    },
    "servers": [
      {
        "url": "https://side-project-vetfolio-manager.vercel.app",
        "description": "Servidor de desarrollo"
      }
    ],
    "paths": {
      "/getPublic": {
        "get": {
          "summary": "Lista todos los casos públicos",
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean"
                      },
                      "patients": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Case"
                        }
                      }
                    }
                  }
                }
              }
            },
            "404": {
                "description": "Error al obtener datos",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
            }
          }
        }
      },
      "/listUser": {
        "get": {
          "summary": "Lista todos los casos privados",
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean"
                      },
                      "patients": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Case"
                        }
                      }
                    }
                  }
                }
              }
            },
            "404": {
                "description": "Error al obtener datos",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
            }
          }
        },
        "delete":  {
            "summary": "Elimina el caso seleccionado",
            "parameters": [
                {
                  "name": "idCase",
                  "in": "query",
                  "required": true,
                  "schema": {
                    "type": "integer"
                  }
                }
              ],
            "responses": {
              "200": {
                "description": "Operación exitosa",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              },
              "404": {
                  "description": "Error al obtener datos",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "success": {
                            "type": "boolean"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
              }
            }
          }
      },
      "/newCase": {
        "post": {
          "summary": "Crea un nuevo caso en usuario registrado",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Caso creado exitosamente",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean"
                      },
                      "patients": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/case": {
        "get": {
          "summary": "Filtra los casos en función de distintos parámetros",
          "parameters": [
            {
              "name": "name",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
                "name": "breed",
                "in": "query",
                "schema": 
                {
                  "type": "string"
                }
              },
              {
                "name": "clinical",
                "in": "query",
                "schema": {
                  "type": "string"
                }
              }
          ],
          "responses": {
            "200": {
                "description": "Operación exitosa",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "message": {
                          "type": "string"
                        },
                        "patients": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Case"
                          }
                        }
                      }
                    }
                  }
                }
              },
            "404": {
              "description": "Error al obtener datos",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Case": {
          "type": "object",
          "properties": {
              "name": "string", 
              "specie": "string", 
              "breed": "string", 
              "gender": "string", 
              "birthday": "integer",
              "clinical": "string", 
              "exploration": "string", 
              "tests": "string",
              "results": "string", 
              "treatment": "string", 
              "evolution": "string", 
              "comments": "string", 
              "public": "integer", 
              "fk_Vet": "integer"
            }
        }
      }
    }
  }